Split the tests into inviscid and euler
